python/requirements.txt
grpcio
grpcio-tools
requests
openweather